+++ templates/simple_counter/Android/app/build.gradle
@@ -4,11 +4,11 @@ plugins {
 }
 
 android {
-    namespace 'com.example.counter'
+    namespace 'com.example.{{workspace}}'
     compileSdk 34
 
     defaultConfig {
-        applicationId "com.example.counter"
+        applicationId "com.example.{{workspace}}"
         minSdk 33
         targetSdk 34
         versionCode 1
@@ -40,7 +40,7 @@ android {
         compose true
     }
     composeOptions {
-        kotlinCompilerExtensionVersion '1.4.7'
+        kotlinCompilerExtensionVersion '1.5.3'
     }
     packagingOptions {
         resources {
@@ -57,7 +57,7 @@ dependencies {
     implementation composeBom
     androidTestImplementation composeBom
 
-    implementation 'androidx.compose.material3:material3:1.2.0-alpha05'
+    implementation 'androidx.compose.material3:material3:1.2.0-alpha06'
 
     // Optional - Integration with activities
     implementation("androidx.activity:activity-compose:1.7.2")
@@ -66,8 +66,8 @@ dependencies {
     // Optional - Integration with LiveData
     implementation("androidx.compose.runtime:runtime-livedata")
 
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.1'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3'
 
     implementation 'androidx.core:core-ktx:1.10.1'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.6.1'
