--- templates/simple_counter/Android/app/src/main/java/com/example/counter/MainActivity.kt
@@ -1,120 +1,0 @@
-@file:OptIn(ExperimentalUnsignedTypes::class)
-
-package com.example.counter
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.*
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.counter.shared.processEvent
-import com.example.counter.shared.view
-import com.example.counter.shared_types.Effect
-import com.example.counter.shared_types.Event
-import com.example.counter.shared_types.Requests
-import com.example.counter.ui.theme.CounterTheme
-import com.example.counter.shared_types.Event as Evt
-import com.example.counter.shared_types.Request as Req
-import com.example.counter.shared_types.ViewModel as MyViewModel
-
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            CounterTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    View()
-                }
-            }
-        }
-    }
-}
-
-sealed class CoreMessage {
-    data class Event(val event: Evt) : CoreMessage()
-}
-
-class Model : ViewModel() {
-    var view: MyViewModel by mutableStateOf(MyViewModel(""))
-        private set
-
-    init {
-        update(CoreMessage.Event(Event.Reset()))
-    }
-
-    fun update(msg: CoreMessage) {
-        val requests: List<Req> =
-            when (msg) {
-                is CoreMessage.Event -> {
-                    Requests.bincodeDeserialize(
-                        processEvent(msg.event.bincodeSerialize())
-                    )
-                }
-            }
-
-        for (req in requests) when (req.effect) {
-            is Effect.Render -> {
-                this.view = MyViewModel.bincodeDeserialize(view())
-            }
-        }
-    }
-}
-
-@Composable
-fun View(model: Model = viewModel()) {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(10.dp),
-    ) {
-        Text(text = model.view.count.toString(), modifier = Modifier.padding(10.dp))
-        Row(horizontalArrangement = Arrangement.spacedBy(10.dp)) {
-            Button(
-                onClick = { model.update(CoreMessage.Event(Event.Reset())) },
-                colors =
-                ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.error
-                )
-            ) { Text(text = "Reset", color = Color.White) }
-            Button(
-                onClick = { model.update(CoreMessage.Event(Event.Increment())) },
-                colors =
-                ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.primary
-                )
-            ) { Text(text = "Increment", color = Color.White) }
-            Button(
-                onClick = { model.update(CoreMessage.Event(Event.Decrement())) },
-                colors =
-                ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.secondary
-                )
-            ) { Text(text = "Decrement", color = Color.White) }
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DefaultPreview() {
-    CounterTheme {
-        View()
-    }
-}
