+++ templates/simple_counter/Android/app/src/main/java/com/example/simple_counter/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.counter.ui.theme
+package com.example.simple_counter.ui.theme
 
 import androidx.compose.ui.graphics.Color
