+++ templates/simple_counter/Android/app/src/main/java/com/example/simple_counter/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.counter.ui.theme
+package com.example.simple_counter.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
