+++ templates/simple_counter/Android/build.gradle
@@ -1,6 +1,6 @@
 plugins {
-    id 'com.android.application' version '8.1.0' apply false
-    id 'com.android.library' version '8.1.0' apply false
-    id 'org.jetbrains.kotlin.android' version '1.8.21' apply false
+    id 'com.android.application' version '8.1.1' apply false
+    id 'com.android.library' version '8.1.1' apply false
+    id 'org.jetbrains.kotlin.android' version '1.9.10' apply false
     id "org.mozilla.rust-android-gradle.rust-android" version "0.9.3" apply false
 }
