+++ templates/simple_counter/Android/shared/build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 android {
-    namespace 'com.example.counter.shared'
+    namespace 'com.example.{{workspace}}.{{core_name}}'
     compileSdk 34
 
     ndkVersion "25.2.9519653"
@@ -42,7 +42,7 @@ android {
 }
 
 dependencies {
-    implementation "net.java.dev.jna:jna:5.12.1@aar"
+    implementation "net.java.dev.jna:jna:5.13.0@aar"
 
     implementation 'androidx.core:core-ktx:1.10.1'
     implementation 'androidx.appcompat:appcompat:1.6.1'
