+++ templates/simple_counter/iOS/SimpleCounter.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
