+++ templates/simple_counter/iOS/SimpleCounter/Info.plist
