+++ templates/simple_counter/iOS/SimpleCounter/core.swift
@@ -1,5 +1,6 @@
 import Foundation
-import SharedTypes
+
+import {{ type_gen_pascal }}
 
 @MainActor
 class Core: ObservableObject {
